+++ processor/src/test/resources/expected/com/example/arez/React4j_ComponentFunctionalInterfaceProp.java
@@ -1,0 +1,97 @@
+package com.example.arez;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.Observable;
+import arez.annotations.ObservableRef;
+import javax.annotation.Generated;
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+import jsinterop.annotations.JsConstructor;
+import jsinterop.annotations.JsPackage;
+import jsinterop.annotations.JsType;
+import jsinterop.base.Js;
+import jsinterop.base.JsPropertyMap;
+import react4j.ComponentConstructorFunction;
+import react4j.NativeAdapterComponent;
+import react4j.ReactConfig;
+
+@ArezComponent(
+    name = "ComponentFunctionalInterfaceProp"
+)
+@Generated("react4j.processor.ReactProcessor")
+abstract class React4j_ComponentFunctionalInterfaceProp extends ComponentFunctionalInterfaceProp {
+  static final ComponentConstructorFunction TYPE = getConstructorFunction();
+
+  @Nonnull
+  private static ComponentConstructorFunction getConstructorFunction() {
+    final ComponentConstructorFunction componentConstructor = NativeReactComponent::new;
+    if ( ReactConfig.enableComponentNames() ) {
+      Js.asPropertyMap( componentConstructor ).set( "displayName", "ComponentFunctionalInterfaceProp" );
+    }
+    return componentConstructor;
+  }
+
+  @Override
+  @Observable(
+      name = "value",
+      expectSetter = false
+  )
+  protected ComponentFunctionalInterfaceProp.TestFunction getValue() {
+    return null != props().getAny( "value" ) ? props().getAny( "value" ).cast() : null;
+  }
+
+  @Nonnull
+  @ObservableRef
+  protected abstract arez.Observable getValueObservable();
+
+  @JsType(
+      isNative = true,
+      namespace = JsPackage.GLOBAL,
+      name = "?"
+  )
+  interface Lifecycle {
+    void componentDidMount();
+
+    void componentDidUpdate(@Nonnull JsPropertyMap<Object> arg0,
+        @Nonnull JsPropertyMap<Object> arg1);
+
+    void componentWillUnmount();
+
+    boolean shouldComponentUpdate(@Nonnull JsPropertyMap<Object> arg0,
+        @Nonnull JsPropertyMap<Object> arg1);
+  }
+
+  private static final class NativeReactComponent extends NativeAdapterComponent<ComponentFunctionalInterfaceProp> implements Lifecycle {
+    @JsConstructor
+    NativeReactComponent(@Nullable final JsPropertyMap<Object> props) {
+      super( props );
+    }
+
+    @Override
+    protected ComponentFunctionalInterfaceProp createComponent() {
+      return new Arez_React4j_ComponentFunctionalInterfaceProp();
+    }
+
+    @Override
+    public void componentDidMount() {
+      performComponentDidMount();
+    }
+
+    @Override
+    public void componentDidUpdate(@Nonnull final JsPropertyMap<Object> arg0,
+        @Nonnull final JsPropertyMap<Object> arg1) {
+      performComponentDidUpdate(arg0,arg1);
+    }
+
+    @Override
+    public void componentWillUnmount() {
+      performComponentWillUnmount();
+    }
+
+    @Override
+    public boolean shouldComponentUpdate(@Nonnull final JsPropertyMap<Object> arg0,
+        @Nonnull final JsPropertyMap<Object> arg1) {
+      return performShouldComponentUpdate(arg0,arg1);
+    }
+  }
+}